Partially fix in memory queries
