+++ package.json
@@ -4,7 +4,7 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "test": "cross-env NODE_ENV=test jest --watchAll --verbose"
+    "test": "NODE_ENV=test jest --watchAll --verbose"
   },
   "keywords": [],
   "author": "",
@@ -19,9 +19,11 @@
     "sqlite3": "^4.1.1"
   },
   "devDependencies": {
+    "alasql": "^0.5.5",
     "chai": "^4.2.0",
     "chai-http": "^4.3.0",
     "jest": "^24.9.0",
-    "mocha": "^6.2.2"
+    "mocha": "^6.2.2",
+    "sqlite": "^3.0.3"
   }
 }
