+++ src/constants/queries.js
@@ -1,5 +1,13 @@
 const QUERIES = {
 	SONGS: {
+		CREATE_TABLE: `
+			CREATE TABLE IF NOT EXISTS songs (
+				id INTEGER PRIMARY KEY AUTOINCREMENT,
+				name VARCHAR(150),
+				price DOUBLE DEFAULT 0.00,
+				genre VARCHAR(150),
+				artist_id INTEGER
+			)`,
 		FIND: 'SELECT * FROM songs WHERE id = ?',
 		LIST: 'SELECT * FROM songs',
 		INSERT: 'INSERT INTO songs (name, price, genre, artist_id) values (?, ?, ?, ?)',
